Added Basic Authentication, used proper routes, and added SSL configuration and https support
Basic Auth:

The main function is the do_auth function, which checks user authentication. For a function to use that, you can just pass in Basic Auth

HTTPS + SSL
I made a cert and key, but removed them from the repo as it is bad information to share a private key. If you need it to run the server, just dm me, and I can send the key and cert. Most of the function are from a third party